- tons of fixes: now all examples and apps build after the port to Eigen (mrpt-performance still doesn't build, but seems a bug in MSVC!) - Deleted class: mrpt::slam::CConsistentObservationAlignment: It implemented the Lu & Milios algorithm, now superseded by graph-slam methods. See the mrpt::graphslam namespace. - mrpt::system::createThread now also has overloaded versions for the thread functions having arguments passed by reference. - BUG FIX: mrpt::utils::net::DNS_resolve_async didn't work properly.

+++ libs/reactivenav/src/CHolonomicND.cpp
@@ -415,7 +415,7 @@ void  CHolonomicND::searchBestGap(
     const int freeSectorsNearTarget = 10;  // 3
     bool theyAreFree = true, caseD1 = false;
     if (TargetSector>(unsigned int)freeSectorsNearTarget &&
-		TargetSector<(obstacles.size()-freeSectorsNearTarget) )
+		TargetSector<(unsigned int)(obstacles.size()-freeSectorsNearTarget) )
     {
         for (int j=-freeSectorsNearTarget;j<=freeSectorsNearTarget;j++)
                 if (obstacles[ TargetSector + j ]<0.95*TargetDist)
